refactor components and add .env to client, server
