+++ client/.env.sample
@@ -1,0 +1,1 @@
+REACT_APP_PASSCODE='777'
