+++ client/public/index.html
@@ -12,5 +12,5 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-  </body>
+  </body>`
 </html>
