+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/App/App.js';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(<App />);
