+++ server/.env.sample
@@ -1,4 +1,9 @@
-USER='userXX'
-PASS='XXpass'
+# client
+
+
+# server
+DBUSER='userXX'
+DBPASS='XXpass'
 DATABASE='dbXX'
-PORT=30XX
+DBPORT=3306
+EXPRESSPORT=30XX
